// admin/models/api.ts
export interface CityResponse {
  id: number;
  name: string;
  isCapital: boolean;
  state: StateResponse;
  createdAt: Date;
  updatedAt: Date;
}

export interface CreateCityRequest {
  name: string;
  isCapital: boolean;
  stateId: number;
}

export interface UpdateCityRequest {
  name: string;
  isCapital: boolean;
  stateId: number;
}
// ...
export interface StateResponse {
  id: number;
  name: string;
  uf: string;
  createdAt: Date;
  updatedAt: Date;
}
// ...
export interface PagedResponse<T> {
  page: number;
  size: number;
  totalElements: number;
  totalPages: number;
  data: T[];
}